moved error functions into utils
