+++ server/controllers/recruiters.js
@@ -1,5 +1,6 @@
 const { User, Recruiter } = require('../db/models/index');
-const { joinConfig, notFoundError, systemError } = require('../utils/db-utils');
+const { joinConfig } = require('../utils/db-utils');
+const { notFoundError, systemError } = require('../utils/utils');
 
 const recruiterControllers = {};
