+++ server/utils/utils.js
@@ -1,3 +1,5 @@
+const createError = require('http-errors');
+
 function normalizePort(val) {
 	const port = parseInt(val, 10);
 
@@ -14,6 +16,21 @@ function normalizePort(val) {
 	return false;
 }
 
+const notFoundError = (type) => {
+	return type ? createError(404, `${type} not found!`) : createError(404);
+};
+
+const systemError = (error) => {
+	return error ? createError(500, error) : createError(500);
+};
+
+const badRequest = (message) => {
+	return message ? createError(400, message) : createError(400);
+};
+
 module.exports = {
-	normalizePort
+	normalizePort,
+	notFoundError,
+	systemError,
+	badRequest
 };
